rename: grouped 'utils', 'contexts' into 'libs'
